% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare.function.R
\name{rezzer}
\alias{rezzer}
\title{A Function To Compare Residual Plots}
\usage{
rezzer(x, y, z)
}
\arguments{
\item{x}{a vector of data}

\item{y}{a vector of numerical data}

\item{z}{a dataframe containing x and y}
}
\description{
This function allows you to compare normal, logged and square rooted residual plots of your data
}
\examples{
rezzer(cyl,mpg,mtcars)
}
\keyword{residuals}

